+++ src/course-home/data/index.js
@@ -1,0 +1,7 @@
+export {
+  fetchDatesTab,
+  fetchOutlineTab,
+  resetDeadlines,
+} from './thunks';
+
+export { reducer } from './slice';
