+++ src/course-home/dates-banner/index.js
@@ -1,0 +1,3 @@
+import DatesBannerContainer from './DatesBannerContainer';
+
+export default DatesBannerContainer;
