+++ src/course-home/dates-tab/index.jsx
@@ -1,0 +1,3 @@
+import DatesTab from './DatesTab';
+
+export default DatesTab;
