+++ src/course-home/outline-tab/index.js
@@ -1,0 +1,1 @@
+export { default } from './OutlineTab';
