+++ src/data/index.js
@@ -1,9 +1,7 @@
 export {
   fetchCourse,
-  fetchDatesTab,
-  fetchOutlineTab,
   fetchSequence,
 } from './thunks';
 
-export { getResumeBlock } from './api';
+export { getResumeBlock, normalizeBlocks } from './api';
 export { reducer } from './slice';
