+++ src/index.scss
@@ -311,5 +311,5 @@ $primary: #1176B2;
 @import 'courseware/course/celebration/CelebrationModal.scss';
 @import 'courseware/course/content-tools/calculator/calculator.scss';
 @import 'courseware/course/content-tools/contentTools.scss';
-@import 'dates-tab/Badge.scss';
-@import 'dates-tab/Day.scss';
+@import 'course-home/dates-tab/Badge.scss';
+@import 'course-home/dates-tab/Day.scss';
