+++ src/tab-page/TabPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
-import { useSelector } from 'react-redux';
 
 import { Header } from '../course-header';
 import { useLogistrationAlert } from '../alerts/logistration-alert';
@@ -11,11 +11,10 @@ import LoadedTabPage from './LoadedTabPage';
 
 function TabPage({
   intl,
+  courseStatus,
   ...passthroughProps
 }) {
   useLogistrationAlert();
-
-  const courseStatus = useSelector(state => state.courseware.courseStatus);
 
   if (courseStatus === 'loading') {
     return (
@@ -47,6 +46,7 @@ function TabPage({
 
 TabPage.propTypes = {
   intl: intlShape.isRequired,
+  courseStatus: PropTypes.string.isRequired,
 };
 
 export default injectIntl(TabPage);
